+++ src/http/client_reader.cpp
@@ -61,7 +61,7 @@ bool ClientReader::put_encoded_data(StreamBuffer encoded){
 			}
 		}
 
-		AUTO(expected, m_queue.cut_off(m_size_expecting));
+		AUTO(expected, m_queue.cut_off(boost::numeric_cast<std::size_t>(m_size_expecting)));
 		if(expecting_new_line){
 			expected.unput(); // '\n'
 			if(expected.back() == '\r'){
@@ -163,7 +163,7 @@ bool ClientReader::put_encoded_data(StreamBuffer encoded){
 
 		case S_IDENTITY:
 			temp64 = std::min<boost::uint64_t>(expected.size(), m_content_length - m_content_offset);
-			on_response_entity(m_content_offset, expected.cut_off(temp64));
+			on_response_entity(m_content_offset, expected.cut_off(boost::numeric_cast<std::size_t>(temp64)));
 			m_content_offset += temp64;
 
 			if(m_content_length == CONTENT_TILL_EOF){
@@ -208,7 +208,7 @@ bool ClientReader::put_encoded_data(StreamBuffer encoded){
 
 		case S_CHUNK_DATA:
 			temp64 = std::min<boost::uint64_t>(expected.size(), m_chunk_size - m_chunk_offset);
-			on_response_entity(m_content_offset, expected.cut_off(temp64));
+			on_response_entity(m_content_offset, expected.cut_off(boost::numeric_cast<std::size_t>(temp64)));
 			m_content_offset += temp64;
 			m_chunk_offset += temp64;
